use ArchiveStatus in Archive.status instead of str
